compiler issues solved

+++ trunk/source/Core/ComponentSystem/BaseComponentContainer.cpp
@@ -66,7 +66,7 @@ namespace GASS
 	ComponentPtr BaseComponentContainer::GetComponent(const std::string &name) const
 	{
 		ComponentPtr comp;
-		for(int i = 0 ; i < m_ComponentVector.size(); i++)
+		for(size_t i = 0 ; i < m_ComponentVector.size(); i++)
 		{
 			if(m_ComponentVector[i]->GetName() == name)
 				return m_ComponentVector[i];
